requirements.txt
reflex==0.8.6
openai
python-dotenv